notification fcmtoken wallet create order edit profile fixed

+++ app/src/main/java/com/teamx/hatlyUser/ui/fragments/hatlymart/stores/model/Doc.kt
@@ -10,7 +10,7 @@ data class Doc(
     val isOpen: Boolean,
     val name: String,
     val owner: String,
-    val ratting: Int,
+    val ratting: Double,
     val totalReviews: Int,
     val type: String
 )
